Adding Python file that writes html for projects
This is a prototype of a python file that writes the HTML code for the DLA list in ASEN. So far, it only works with a specific .csv file (file.csv) and the ASEN major. This will have to be changed later. Also, do note that the .csv file it reads must be in a rather specific format. Look at file.csv for reference.